test: created new tests and restructured tests to match standards
Todo: More tests needs to be added for some functions, Se TODOs
